requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
importlib-metadata==6.8.0
Markdown==3.5.1
pip==23.3.1
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
